split basic_ops.py
