+++ src/bequem/nodes/basic_operations/compute_projection.py
@@ -1,0 +1,39 @@
+from __future__ import annotations
+import numpy as np
+
+from bequem.circuit import Circuit
+from bequem.subspace.subspace import Subspace
+from bequem.nodes.node import Node
+
+
+class ComputeProjection(Node):
+    def __init__(self, subspace: Subspace):
+        self.subspace = subspace
+
+    def children(self) -> list[Node]:
+        return []
+
+    def _subspace_in(self) -> Subspace:
+        # TODO: most of the extra zeros are actually ancillas
+        return Subspace(
+            self.subspace.registers,
+            self.circuit.tq_circuit.n_qubits - self.subspace.total_qubits,
+        )
+
+    def _subspace_out(self) -> Subspace:
+        return Subspace(
+            self.subspace.registers,
+            self.circuit.tq_circuit.n_qubits - self.subspace.total_qubits,
+        )
+
+    def _normalization(self) -> float:
+        return 1
+
+    def compute(self, input: np.ndarray) -> np.ndarray:
+        return input
+
+    def compute_adjoint(self, input: np.ndarray) -> np.ndarray:
+        return input
+
+    def _circuit(self) -> Circuit:
+        return self.subspace.circuit()
